Rename factories to have related function in its name
