+++ packages/resizin/src/buildUrl/index.ts
@@ -1,2 +1,2 @@
 export { default } from './buildUrl';
-export { default as buildUrlFactory } from './factory';
+export { default as buildUrlFactory } from './buildUrlFactory';
